chore(controller): renamed invited admin controller file
